What - read function is ready now; replace log_path with writable_tmp_path, we need store log and cookie to this path, and find the root cause of liboauth nonce is not correct; and add download by range, now it could use read function.
@Description/root cause - Why
1. we need a path to store both log and cookie for curl.
2. oauth nonce is not correct under FUSE mode, because in main function we
call oauth once, it seems we could not do that. we need to move
kpfs_get_root_path() to fuse operations, then oauth nonce works well.
3. this oauth nonce bug is found by run the FUSE hello.c sample with oauth, it
works well if put oauth_nonce in fuse operations, but it went abnormal if I
put oauth_nonce in main function.

@Resolution - How
1. add download by range unit test
2. move kpfs_get_root_path() oauth related code to fuse operations, we should
not call it in main function.
3. change log_path to writable_tmp_path.
4. add KPFS_APP_NAME, it will used in curl user agent.
5. add cookie.txt for curl cookie temp file storing.
6. change write_func of curl to use a structure, we need to get count of buf,
and return to caller. 

@Reviewer
None

@Tickets
NONE